+++ sim/druid/moonfire.go
@@ -13,9 +13,9 @@ const MoonfireRanks = 10
 var MoonfireSpellId = [MoonfireRanks + 1]int32{0, 8921, 8924, 8925, 8926, 8927, 8928, 8929, 9833, 9834, 9835}
 var MoonfiresSpellCoeff = [MoonfireRanks + 1]float64{0, .06, .094, .128, .15, .15, .15, .15, .15, .15, .15}
 var MoonfiresSellDotCoeff = [MoonfireRanks + 1]float64{0, .052, .081, .111, .13, .13, .13, .13, .13, .13, .13}
-var MoonfireBaseDamage = [MoonfireRanks + 1][]float64{{0}, {9, 12}, {17, 21}, {30, 37}, {44, 53}, {70, 82}, {91, 108}, {117, 137}, {143, 168}, {172, 200}, {195, 228}}
-var MoonfireBaseDotDamage = [MoonfireRanks + 1]float64{0, 12, 32, 52, 80, 124, 164, 212, 264, 320, 384}
-var MoonfireDotTicks = [MoonfireRanks + 1]int32{0, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4}
+var MoonfireBaseDamage = [MoonfireRanks + 1][]float64{{0}, {7, 9}, {17, 21}, {30, 37}, {46, 54}, {70, 82}, {91, 108}, {117, 137}, {143, 168}, {172, 200}, {195, 228}}
+var MoonfireBaseDotDamage = [MoonfireRanks + 1]float64{0, 16, 32, 52, 80, 124, 164, 212, 264, 320, 384}
+var MoonfireDotTicks = [MoonfireRanks + 1]int32{0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4}
 var MoonfireManaCost = [MoonfireRanks + 1]float64{0, 25, 50, 75, 105, 150, 190, 235, 280, 325, 375}
 var MoonfireLevel = [MoonfireRanks + 1]int{0, 4, 10, 16, 22, 28, 34, 40, 46, 52, 58}
 
@@ -51,7 +51,7 @@ func (druid *Druid) getMoonfireBaseConfig(rank int) core.SpellConfig {
 	spellDotCoeff := MoonfiresSellDotCoeff[rank]
 	baseDamageLow := MoonfireBaseDamage[rank][0]
 	baseDamageHigh := MoonfireBaseDamage[rank][1]
-	baseDotDamage := (MoonfireBaseDotDamage[rank] / float64(ticks))
+	baseDotDamage := MoonfireBaseDotDamage[rank] / float64(MoonfireDotTicks[rank]) // Power of Nature adds ticks, not spreads them
 	manaCost := MoonfireManaCost[rank]
 	level := MoonfireLevel[rank]